Added Camera Orientation Mode Strings
I've recently added the ability for the camera to auto-detect the held orientation to let users take photos in "portrait" orientation if they wish. In addition to this, there is a UI option to have the orientation chosen automatically, or manually by the user - these strings are for that option.