EN - tiny fixes
Small things I noticed while working on the FR translation. The most important one is what I think is an inversion of "increase" and "lower". I'm 95% sure that if the setting refers to a length of time then lowering the time between clicks would be done if double clicking is too hard, not too easy.