Espaces insécables
French is an odd language.

We are supposed to insert spaces before each punctuation that is composed of 2 parts, such as ":", "?" or "!".
This is weird enough.

But what’s even weirder is that we are not just supposed to use plain old regular spaces.
Oh no, we need to use no-break spaces, to indicate "yes, put a space before that question mark, but UNDER NO CIRCUMSTANCES shall this question mark be separated from its preceding word with a line break".
And to top it off, we have both full-width no-break spaces which must be used before ":", and narrow no-break spaces which must be used before "?" and "!".

Needless to say, 99.9% of the French population has never manually typed a no-break space in their life.
But I am a stickler for typography, so I am righting the wrong in the French locale of Resonite!

I have tested that the no-break spaces display properly in game, in my testing there did not seem to be any visual issue - either on release or pre-release.